Always modify instead of creating new when annotation type J.FieldAccess
Annotation type is J.FieldAccess anytime the annotation is fully qualified.
The code in org.openrewrite.java.testing.junit5.UpdateTestAnnotation
(which this is based on) would build a new annotation instance instead of
modifying the type of the existing annotation. It turns out the latter also
works fine for this recipe's scope.